added 2 templates and split packages
