+++ src/main/java/com/zakharuk/quickdr/entity/Doctor.java
@@ -1,4 +1,4 @@
-package com.zakharuk.quickdr;
+package com.zakharuk.quickdr.entity;
 
 import javafx.util.Pair;
